backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
supabase==2.5.0
pandas==2.2.2
numpy==1.26.4
statsmodels==0.14.2
scipy==1.13.1
python-dotenv==1.0.1
pydantic==2.7.4
python-multipart==0.0.9
httpx==0.27.0
